Resolve conversion from ssize_t to int for readlink return value

+++ Compat.h
@@ -50,10 +50,10 @@ int Compat_openat(openat_arg_t dirpath, const char* pathname, int flags);
 
 #endif /* HAVE_OPENAT */
 
-int Compat_readlinkat(int dirfd,
-                   const char* dirpath,
-                   const char* pathname,
-                   char* buf,
-                   size_t bufsize);
+ssize_t Compat_readlinkat(int dirfd,
+                          const char* dirpath,
+                          const char* pathname,
+                          char* buf,
+                          size_t bufsize);
 
 #endif /* HEADER_Compat */
